refactor: refactored index.ts for web and moved types to src
