+++ packages/react-native-audio-api/src/index.native.ts
@@ -17,6 +17,7 @@ export { default as BiquadFilterNode } from './core/BiquadFilterNode';
 export { default as GainNode } from './core/GainNode';
 export { default as OscillatorNode } from './core/OscillatorNode';
 export { default as StereoPannerNode } from './core/StereoPannerNode';
+
 export {
   OscillatorType,
   BiquadFilterType,
@@ -24,4 +25,12 @@ export {
   ChannelInterpretation,
   ContextState,
   WindowType,
-} from './core/types';
+  PeriodicWaveConstraints,
+} from './types';
+
+export {
+  IndexSizeError,
+  InvalidAccessError,
+  InvalidStateError,
+  RangeError,
+} from './errors';
