+++ packages/react-native-audio-api/src/web-core/PeriodicWave.tsx
@@ -1,0 +1,8 @@
+export default class PeriodicWave {
+  /** @internal */
+  readonly periodicWave: globalThis.PeriodicWave;
+
+  constructor(periodicWave: globalThis.PeriodicWave) {
+    this.periodicWave = periodicWave;
+  }
+}
